fix eslint errors and console errors
