+++ src/app.js
@@ -1,33 +1,38 @@
 import tasks from './tasks.js';
-import renderTasks from './index.js';
-console.log("app!")
-export default function addTask(description) {
-  console.log("add!")
+
+export function addTask(tasks, description) {
+  console.log("add")
   const task = {
     description,
     completed: false,
     index: tasks.length + 1,
   };
-  console.log('hi');
   tasks.push(task);
   localStorage.setItem('tasks', JSON.stringify(tasks));
-  renderTasks();
 }
 
-const newTaskForm = document.querySelector('form');
-newTaskForm.addEventListener('submit', (event) => {
-  event.preventDefault();
-  const newTaskInput = newTaskForm.querySelector('.new-task-input');
-  addTask(newTaskInput.value);
-  newTaskInput.value = '';
-});
+export function editTask(index) {
+  console.log("edit")
+  const taskItem = document.querySelector(`#todo-list li:nth-child(${index + 1})`);
+  const descriptionElement = taskItem.querySelector('p');
+  const editInput = document.createElement('input');
+  editInput.type = 'text';
+  editInput.value = descriptionElement.innerText;
+  editInput.addEventListener('blur', () => {
+    tasks[index].description = editInput.value;
+    localStorage.setItem('tasks', JSON.stringify(tasks));
+  });
+  descriptionElement.replaceWith(editInput);
+  editInput.focus();
+}
 
-const clearCompletedBtn = document.querySelector('.clear-completed-btn');
-clearCompletedBtn.addEventListener('click', () => {
-  tasks = tasks.filter((task) => !task.completed);
+export function deleteTask(index) {
+  console.log("delete")
+  tasks.splice(index, 1);
+  // Update the indexes of the remaining tasks
   tasks.forEach((task, i) => {
     task.index = i + 1;
+    return task;
   });
   localStorage.setItem('tasks', JSON.stringify(tasks));
-  renderTasks();
-});
+}
