+++ src/index.html
@@ -15,7 +15,7 @@
           <h1>Today's To Do</h1>
           <i class="fa">&#xf021;</i>
         </header>
-        <form id="todo-form">
+        <form>
           <input type="text" class="new-task-input" placeholder="Add to your list..." />
           <button type="submit" class="add-task-btn"><p>&#x21B2;</p></button>
         </form>
@@ -28,6 +28,5 @@
 
     </div>
 
-    <script type="module" src="./index.js"></script>
   </body>
 </html>
