+++ src/index.js
@@ -1,12 +1,13 @@
 import tasks from './tasks.js';
 import './style.css';
 import './storage.js';
-console.log("index!")
+import { addTask, editTask, deleteTask } from './app.js';
+
 export default function renderTasks() {
+  console.log("render")
   const todoList = document.getElementById('todo-list');
   todoList.innerHTML = ''; // clear previous items
   tasks.sort((a, b) => a.index - b.index);
-  console.log("render!");
   tasks.forEach((task, index) => {
     const taskItem = document.createElement('li');
     taskItem.innerHTML += `<input class='checkbox' type="checkbox" ${task.completed ? 'checked' : ''}/><p>${task.description}</p> <i class="fa">&#xf142;</i>`;
@@ -15,27 +16,12 @@ export default function renderTasks() {
       localStorage.setItem('tasks', JSON.stringify(tasks));
       renderTasks();
     });
-    taskItem.querySelector('p').addEventListener('click', function editTask() {
-      const descriptionElement = this;
-      const editInput = document.createElement('input');
-      editInput.type = 'text';
-      editInput.value = descriptionElement.innerText;
-      editInput.addEventListener('blur', () => {
-        tasks[index].description = editInput.value;
-        localStorage.setItem('tasks', JSON.stringify(tasks));
-        renderTasks();
-      });
-      descriptionElement.replaceWith(editInput);
-      editInput.focus();
+    taskItem.querySelector('p').addEventListener('click', () => {
+      editTask(index);
+      renderTasks();
     });
     taskItem.querySelector('i').addEventListener('click', () => {
-      // deleteTasks
-      const index = tasks.findIndex((task) => task === taskItem);
-      tasks.splice(index, 1);
-      tasks.forEach((task, i) => {
-        task.index = i + 1;
-      });
-      localStorage.setItem('tasks', JSON.stringify(tasks));
+      deleteTask(index);
       renderTasks();
     });
 
@@ -46,3 +32,21 @@ export default function renderTasks() {
     todoList.appendChild(taskItem);
   });
 }
+
+const newTaskForm = document.querySelector('form');
+newTaskForm.addEventListener('submit', (event) => {
+  event.preventDefault();
+  const newTaskInput = document.querySelector('.new-task-input');
+  addTask(tasks, newTaskInput.value);
+  newTaskInput.value = '';
+});
+
+const clearCompletedBtn = document.querySelector('.clear-completed-btn');
+clearCompletedBtn.addEventListener('click', () => {
+  tasks = tasks.filter((task) => !task.completed);
+  tasks.forEach((task, i) => {
+    task.index = i + 1;
+  });
+  localStorage.setItem('tasks', JSON.stringify(tasks));
+  renderTasks();
+});
