+++ webpack.config.js
@@ -29,15 +29,15 @@ module.exports = {
         use: {
           loader: 'babel-loader',
           options: {
-            presets: ['@babel/preset-env']
-          }
-        }
+            presets: ['@babel/preset-env'],
+          },
+        },
       },
       {
         test: /\.css$/i,
         use: ['style-loader', 'css-loader'],
-      }
-    ]
-  }
-  
+      },
+    ],
+  },
+
 };
